LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/generic-0_10_1.sub-expected.txt

implicit port number matching fails with a different port


PASS Prevents access to external scripts.
FAIL Should fire violation events for every failed violation assert_equals: expected "script-src-elem" but got "script-src"
